+++ CLibrary/WebContent/css/index.css
@@ -24,11 +24,15 @@ a{
 
 .bookContent{
 	padding: 0;
-	width: 355px;
-    margin-right: 15px;
+	width: 257px;
+    margin-right: 14px;
 }
 
 .bookTitle{
 	background-color: rgba(204,232,255,0.85);
 }
 
+img{
+	width: 257px;
+	height: 300px;
+}
