+++ CLibrary/src/servlet/KuramotoServlet.java
@@ -14,7 +14,6 @@ import javax.servlet.http.HttpSession;
 
 import dto.BooksDTO;
 import dto.ForListDTO;
-import dto.StaffsDTO;
 import model.KuramotoDAO;
 
 /**
@@ -39,42 +38,42 @@ public class KuramotoServlet extends HttpServlet {
 			throws ServletException, IOException {
 
 		request.setCharacterEncoding("UTF-8");//文字コードを指定
-
+		HttpSession session = request.getSession();
 		//**********************************************************
 		//全従業員一覧表示テスト
 		//**********************************************************
 
-		//ＤＡＯのインスタンスを取得
-		KuramotoDAO kdao1 = new KuramotoDAO();
+//		//ＤＡＯのインスタンスを取得
+//		KuramotoDAO kdao1 = new KuramotoDAO();
+//
+//		//ArrayListを作成して、クリア処理
+//		List<StaffsDTO> staffsListAll = new ArrayList<>();
+//		staffsListAll.clear();
+//
+//		//メソッドallStaffsList()を実行
+//		staffsListAll = kdao1.allStaffsList();
+//
+//		//全従業員の一覧をセッションスコープに保存
+//		HttpSession session = request.getSession();
+//		session.setAttribute("staffsListAll", staffsListAll);
 
-		//ArrayListを作成して、クリア処理
-		List<StaffsDTO> staffsListAll = new ArrayList<>();
-		staffsListAll.clear();
-
-		//メソッドallStaffsList()を実行
-		staffsListAll = kdao1.allStaffsList();
-
-		//全従業員の一覧をセッションスコープに保存
-		HttpSession session = request.getSession();
-		session.setAttribute("staffsListAll", staffsListAll);
-
-		//**********************************************************
-		//全本の一覧表示テスト
-		//**********************************************************
-
-		//ＤＡＯのインスタンスを取得
-		KuramotoDAO kdao2 = new KuramotoDAO();
-
-		//ArrayListを作成して、クリア処理
-		List<BooksDTO> booksListAll = new ArrayList<>();
-		booksListAll.clear();
-
-		//メソッドallStaffsList()を実行
-		booksListAll = kdao2.allBooksList();
-
-		//全従業員の一覧をセッションスコープに保存
-		//HttpSession session = request.getSession();
-		session.setAttribute("booksListAll", booksListAll);
+//		//**********************************************************
+//		//全本の一覧表示テスト
+//		//**********************************************************
+//
+//		//ＤＡＯのインスタンスを取得
+//		KuramotoDAO kdao2 = new KuramotoDAO();
+//
+//		//ArrayListを作成して、クリア処理
+//		List<BooksDTO> booksListAll = new ArrayList<>();
+//		booksListAll.clear();
+//
+//		//メソッドallStaffsList()を実行
+//		booksListAll = kdao2.allBooksList();
+//
+//		//全従業員の一覧をセッションスコープに保存
+//		//HttpSession session = request.getSession();
+//		session.setAttribute("booksListAll", booksListAll);
 
 		//**********************************************************
 		//★人気の本ランキングＴＯＰ3
@@ -93,39 +92,39 @@ public class KuramotoServlet extends HttpServlet {
 		//HttpSession session = request.getSession();
 		session.setAttribute("top3List", top3List);
 
-		//**********************************************************
-		//全貸出履歴リスト
-		//**********************************************************
-		//ＤＡＯのインスタンスを取得
-		KuramotoDAO kdao4 = new KuramotoDAO();
+//		//**********************************************************
+//		//全貸出履歴リスト
+//		//**********************************************************
+//		//ＤＡＯのインスタンスを取得
+//		KuramotoDAO kdao4 = new KuramotoDAO();
+//
+//		//ArrayListを作成して、クリア処理
+//		List<ForListDTO> rentlogsAll = new ArrayList<>();
+//		rentlogsAll.clear();
+//
+//		//メソッドallStaffsList()を実行
+//		rentlogsAll = kdao4.rentLogsAllList();
+//
+//		//全従業員の一覧をセッションスコープに保存
+//		//HttpSession session = request.getSession();
+//		session.setAttribute("rentlogsAll", rentlogsAll);
 
-		//ArrayListを作成して、クリア処理
-		List<ForListDTO> rentlogsAll = new ArrayList<>();
-		rentlogsAll.clear();
-
-		//メソッドallStaffsList()を実行
-		rentlogsAll = kdao4.rentLogsAllList();
-
-		//全従業員の一覧をセッションスコープに保存
-		//HttpSession session = request.getSession();
-		session.setAttribute("rentlogsAll", rentlogsAll);
-
-		//**********************************************************
-		//全貸出中リスト
-		//**********************************************************
-		//ＤＡＯのインスタンスを取得
-		KuramotoDAO kdao5 = new KuramotoDAO();
-
-		//ArrayListを作成して、クリア処理
-		List<ForListDTO> rentNowAll = new ArrayList<>();
-		rentNowAll.clear();
-
-		//メソッドallStaffsList()を実行
-		rentNowAll = kdao5.rentNowAllList();
-
-		//全従業員の一覧をセッションスコープに保存
-		//HttpSession session = request.getSession();
-		session.setAttribute("rentNowAll", rentNowAll);
+//		//**********************************************************
+//		//全貸出中リスト
+//		//**********************************************************
+//		//ＤＡＯのインスタンスを取得
+//		KuramotoDAO kdao5 = new KuramotoDAO();
+//
+//		//ArrayListを作成して、クリア処理
+//		List<ForListDTO> rentNowAll = new ArrayList<>();
+//		rentNowAll.clear();
+//
+//		//メソッドallStaffsList()を実行
+//		rentNowAll = kdao5.rentNowAllList();
+//
+//		//全従業員の一覧をセッションスコープに保存
+//		//HttpSession session = request.getSession();
+//		session.setAttribute("rentNowAll", rentNowAll);
 
 		//**********************************************************
 		//２週間以上借りているリスト
